Attempt to handle LARS regression noise failures
cvx is giving inexplicable error when called from
update_spatial_components...which is very rare,, only observe in a
handful of datasets! Can try to use default, most stable solver, but
also embed LARS in try-catch, and set pixel to 0 weights if LARS fails
for that pixel. This is important particularly for computers without cvx
installed.